+++ designer/deployment-manager-api/src/test/scala/pl/touk/nussknacker/engine/api/deployment/cache/CachingProcessStateDeploymentManagerSpec.scala
@@ -33,6 +33,7 @@ class CachingProcessStateDeploymentManagerSpec
       NoDeploymentSynchronisationSupport,
       NoDeploymentsStatusesQueryForAllScenariosSupport,
       NoSchedulingSupport,
+      NoLiveDataPreviewSupport,
     )
 
     val results = List(
@@ -53,6 +54,7 @@ class CachingProcessStateDeploymentManagerSpec
       NoDeploymentSynchronisationSupport,
       NoDeploymentsStatusesQueryForAllScenariosSupport,
       NoSchedulingSupport,
+      NoLiveDataPreviewSupport,
     )
 
     val firstInvocation = cachingManager.getProcessStatesDeploymentIdNow(DataFreshnessPolicy.CanBeCached)
@@ -72,6 +74,7 @@ class CachingProcessStateDeploymentManagerSpec
       NoDeploymentSynchronisationSupport,
       NoDeploymentsStatusesQueryForAllScenariosSupport,
       NoSchedulingSupport,
+      NoLiveDataPreviewSupport,
     )
 
     val resultForFresh = cachingManager.getProcessStatesDeploymentIdNow(DataFreshnessPolicy.Fresh)