+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/runner/FlinkScenarioJob.scala
@@ -4,7 +4,7 @@ import org.apache.flink.api.common.JobExecutionResult
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
 import pl.touk.nussknacker.engine.{BaseModelData, ModelData}
 import pl.touk.nussknacker.engine.ModelData.BaseModelDataExt
-import pl.touk.nussknacker.engine.api.ProcessVersion
+import pl.touk.nussknacker.engine.api.{ProcessListener, ProcessVersion}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.DeploymentData
 import pl.touk.nussknacker.engine.process.{ExecutionConfigPreparer, FlinkJobConfig}
@@ -18,13 +18,15 @@ object FlinkScenarioJob {
       scenario: CanonicalProcess,
       processVersion: ProcessVersion,
       deploymentData: DeploymentData,
-      env: StreamExecutionEnvironment
+      env: StreamExecutionEnvironment,
+      processListeners: List[ProcessListener],
   ): JobExecutionResult =
     new FlinkScenarioJob(modelData.asInvokableModelData).run(
       scenario,
       processVersion,
       deploymentData,
-      env
+      env,
+      processListeners,
     )
 
 }
@@ -35,9 +37,10 @@ class FlinkScenarioJob(modelData: ModelData) {
       scenario: CanonicalProcess,
       processVersion: ProcessVersion,
       deploymentData: DeploymentData,
-      env: StreamExecutionEnvironment
+      env: StreamExecutionEnvironment,
+      processListeners: List[ProcessListener],
   ): JobExecutionResult = {
-    val compilerFactory         = new FlinkProcessCompilerDataFactory(modelData, deploymentData)
+    val compilerFactory         = new FlinkProcessCompilerDataFactory(modelData, deploymentData, processListeners)
     val executionConfigPreparer = ExecutionConfigPreparer.defaultChain(modelData)
     val registrar =
       FlinkProcessRegistrar(compilerFactory, FlinkJobConfig.parse(modelData.modelConfig), executionConfigPreparer)