+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkDeploymentManager.scala
@@ -382,6 +382,9 @@ class FlinkDeploymentManager(
       .map(new FlinkScenarioCompilationDependencies(_))
   }
 
+  override def liveDataPreviewSupport: LiveDataPreviewSupport =
+    jobRunner.liveDataPreviewSupport
+
 }
 
 object FlinkDeploymentManager {